refactor(app): extract tab manager
